+++ src/components/Calculator.tsx
@@ -24,8 +24,8 @@ import DateRangeTwoToneIcon from "@material-ui/icons/DateRangeTwoTone";
 import PowerIcon from "@material-ui/icons/Power";
 import MonetizationOnIcon from "@material-ui/icons/MonetizationOn";
 import InsertDriveFileIcon from "@material-ui/icons/InsertDriveFile";
-import { VecRecord } from "../model/vec-csv";
 import { max, median, q25, q75, quantile } from "../model/maths-helper";
+import { VecRecord } from "../model/vecRecord";
 
 const useStyles = makeStyles((theme) => ({
   container: {
